fix(markee): harden bridge claims
