+++ pkg/contracts/config/networks.json
@@ -1161,7 +1161,7 @@
         "PAUSE_CONTROLLER": "0xb32991c77ff0a57a1fb04b76bba03970198dcdbf",
         "STREAMING_ESCROW_FACTORY": "0x3b3333665de09494c2491171c06a6a161f319032",
         "MARKEE_ROUTER": "0xa9d9b3a9CE7edA7000A4baD8Af0863F716eA9d30",
-        "MARKEE_LIFI_BRIDGE_ADAPTER": "0xA6A5225Cdcd8697E341E9952C0744cfFd6296251"
+        "MARKEE_LIFI_BRIDGE_ADAPTER": "0x0B62e2d989d30cbdbB14287e4C048560c8Adf0Ca"
       },
       "PROXIES": {
         "REGISTRY_FACTORY": "0xc93830dd463516ed5f28f6cd4f837173b87ff389",
